# apidocs/swagger.yaml
basePath: /
definitions:
  github_com_stockflow_server_internal_response.Envelope:
    properties:
      code:
        description: 成功 0（数字）；失败为模块命名空间错误码字符串（COMMON_* 等）
      data: {}
      details:
        description: 校验失败必带（api.md §4）
      message:
        type: string
      request_id:
        type: string
    type: object
...
  internal_sysops.jobStatusPayload:
    properties:
      enabled:
        type: boolean
    type: object
  internal_sysops.saveConfigsPayload:
    properties:
      items:
        items:
          properties:
            key:
              type: string
            value:
              type: string
          type: object
        type: array
    type: object
...
info:
  contact: {}
  description: StockFlow（库流智能仓储管理系统）后端 API。统一信封 {code,message,data,request_id,details}（api.md
    §2）。
  title: StockFlow API
  version: "1.0"
paths:
...
  /api/logs/logins:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/logs/logins
      tags:
      - 系统运维
  /api/logs/operations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/logs/operations
      tags:
      - 系统运维
  /api/logs/operations/{id}:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/logs/operations/:id
      tags:
      - 系统运维
  /api/notifications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/notifications
      tags:
      - 系统运维
  /api/notifications/{id}/read:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/notifications/:id/read
      tags:
      - 系统运维
  /api/notifications/read-all:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/notifications/read-all
      tags:
      - 系统运维
  /api/notifications/unread-count:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/notifications/unread-count
      tags:
      - 系统运维
...
  /api/system/backups:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/system/backups
      tags:
      - 系统运维
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/system/backups
      tags:
      - 系统运维
  /api/system/backups/{id}/download:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/system/backups/:id/download
      tags:
      - 系统运维
  /api/system/backups/pg-dump-template:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/system/backups/pg-dump-template
      tags:
      - 系统运维
  /api/system/configs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/system/configs
      tags:
      - 系统运维
    put:
      consumes:
      - application/json
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_sysops.saveConfigsPayload'
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/system/configs
      tags:
      - 系统运维
  /api/system/jobs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/system/jobs
      tags:
      - 系统运维
  /api/system/jobs/{id}/run-logs:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/system/jobs/:id/run-logs
      tags:
      - 系统运维
  /api/system/jobs/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_sysops.jobStatusPayload'
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: PUT /api/system/jobs/:id/status
      tags:
      - 系统运维
  /api/system/monitor:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/system/monitor
      tags:
      - 系统运维
...
securityDefinitions:
  BearerAuth:
    description: 值格式 "Bearer <access_token>"（api.md §6.1）。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"